chore: move option to same area as follow after reply option
